requirements.txt
apturl==0.5.2
asn1crypto==0.24.0
Babel==2.6.0
Brlapi==0.6.6
certifi==2019.3.9
chardet==3.0.4
command-not-found==0.3
cryptography==2.1.4
cupshelpers==1.0
defer==1.0.6
distro-info==0.18
Django==2.2
django-ckeditor==5.6.1
django-js-asset==1.2.2
django-model-utils==3.1.2
django-phone-field==1.7.0
django-phonenumber-field==2.3.1
django-taggit==1.1.0
docopt==0.6.2
httplib2==0.9.2
idna==2.8
keyring==10.6.0
keyrings.alt==3.0
language-selector==0.1
launchpadlib==1.10.6
lazr.restfulclient==0.13.5
lazr.uri==1.0.3
louis==3.5.0
macaroonbakery==1.1.3
Mako==1.0.7
MarkupSafe==1.0
mysqlclient==1.4.2.post1
oauth==1.0.1
olefile==0.45.1
pexpect==4.2.1
phonenumbers==8.10.9
Pillow==5.1.0
pipreqs==0.4.9
protobuf==3.0.0
pycairo==1.16.2
pycrypto==2.6.1
pycups==1.9.73
pygobject==3.26.1
pymacaroons==0.13.0
PyMySQL==0.9.3
PyNaCl==1.1.2
pyRFC3339==1.0
python-apt==1.6.2
python-debian==0.1.32
pytz==2019.1
pyxdg==0.25
PyYAML==3.12
reportlab==3.4.0
requests==2.21.0
requests-unixsocket==0.1.5
SecretStorage==2.3.1
simplejson==3.13.2
six==1.11.0
sqlparse==0.3.0
system-service==0.3
systemd-python==234
ubuntu-drivers-common==0.0.0
ufw==0.35
unattended-upgrades==0.1
urllib3==1.24.1
usb-creator==0.3.3
wadllib==1.3.2
xkit==0.0.0
yarg==0.1.9
zope.interface==4.3.2
